Refactoring the conneg routines to their own helper module
